+++ launcher_common/include/launcher_common/VideoDeviceInfoProvider.h
@@ -3,14 +3,18 @@
 #include <set>
 #include <vector>
 #include <string>
+#include <windows.h>
+
+#define USE_D3D9 1
 
 // Only include D3D header if one has not been included before (fix for afx.h including d3d9 and DF using d3d8)
-#ifndef DIRECT3D_VERSION
-#include <windows.h>
+#if USE_D3D9
+#include <d3d9.h>
+#elif !defined(DIRECT3D_VERSION)
 #include <d3d8.h>
-#endif
-
+#else
 struct IDirect3D8;
+#endif
 
 class VideoDeviceInfoProvider
 {
@@ -32,6 +36,10 @@ public:
     bool has_anisotropy_support(unsigned adapter);
 
 private:
+#if USE_D3D9
+    IDirect3D9* m_d3d;
+#else
     IDirect3D8* m_d3d;
+#endif
     HMODULE m_lib;
 };